day15b
Though a very slow solution: loop through all the rows and look for the
first position where a beacon could fit.

This uncovered some mistakes in the solution to the first part of the
problem as well: should not construct impossible intervals, and when
merging, since they're all on integer coordinates, we should also merge
intervals with endpoints next to each other.